refactor(skills): relocate SkillBarPicker to neutral src/components/skills (B3a)
Lift the presentational SkillBarPicker (front/back bar editor + skill picker
dialog) out of build-editor into the neutral `src/components/skills` so
non-build-editor surfaces -- the Loadout Manager (B3b adoption) and the roster
panels -- can reuse it without a `loadout-manager -> build-editor` component
cycle.

- Add `src/components/skills/SkillBarPicker.tsx`; leave a thin @deprecated
  re-export shim at the old pickers path so existing import sites keep working.
- Move the SkillBarPicker test suite to `src/components/skills/__tests__/`.
- Repoint the two direct consumers (build-editor SkillsSection, roster
  SlotFullModePanel) at the neutral path.

The component's only build-editor dependencies are leaf modules -- data
(esoStaticData), theme (classColorMap), the setupTransfer util, and the
PickerDialog primitive -- none of which import back into components, so no
cycle forms (same reasoning the gear/engine relocations use).

Rebased onto current main: replays the relocation over main's latest
SkillBarPicker rewrite (which added the PickerDialog/setupTransfer
dependencies since the branch was cut), so no behavior change and main's
recent skill-picker work is preserved. The GearPicker test suites main added
under pickers/__tests__ are left in place.
