+++ src/features/build-editor/components/sections/SkillsSection.tsx
@@ -8,9 +8,10 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
+import { SkillBarPicker } from '@/components/skills/SkillBarPicker';
+
 import { selectActiveSetup, selectBuildClassSkillLines } from '../../store/buildEditorSelectors';
 import { setSkills } from '../../store/buildEditorSlice';
-import { SkillBarPicker } from '../pickers/SkillBarPicker';
 
 const SkillsSectionComponent: React.FC = () => {
   const dispatch = useDispatch();
